+++ tests/test-add-admin-css.php
@@ -138,11 +138,11 @@ class Add_Admin_CSS_Test extends WP_UnitTestCase {
 	}
 
 	public function test_plugin_framework_class_name() {
-		$this->assertTrue( class_exists( 'c2c_AddAdminCSS_Plugin_046' ) );
+		$this->assertTrue( class_exists( 'c2c_AddAdminCSS_Plugin_049' ) );
 	}
 
 	public function test_plugin_framework_version() {
-		$this->assertEquals( '046', c2c_AddAdminCSS::instance()->c2c_plugin_version() );
+		$this->assertEquals( '049', c2c_AddAdminCSS::instance()->c2c_plugin_version() );
 	}
 
 	public function test_version() {